# Hash#store is faster than delete then []=