@schneems 2016-04-13
We can keep `Hash#[]=` it is the exact same method as `Hash#store`. It looks like `Hash#delete` was added to keep track of whether or not the key previously existed in https://github.com/rails/sprockets/commit/a18088d7d49e64ec197bbeb8f53d2f13ed8777fe#diff-ff5a66a90c18abb318daf9c2b3164370R21 but it's not being used now, so we can safely get rid of it.