@BlakeMesdag 2016-04-13
@schneems the benchmarks I ran actually indicated it was slower to use `[]=` in all cases.

This is likely due to the way Ruby expands the arguments when presented in the two arg form `hash[key] = value`, which requires parsing args on both the left, and right of the assignment operator. This is my assumption since the methods are assigned to the same C function underneath: https://github.com/ruby/ruby/blob/trunk/hash.c#L4284-L4285