@schneems 2016-04-13
Sorry that was the wrong link, the delete was in there from the beginning, not sure why https://github.com/rails/sprockets/commit/a18088d7d49e64ec197bbeb8f53d2f13ed8777fe#diff-ff5a66a90c18abb318daf9c2b3164370R30.

Store isn't faster, it's about the same

```
require 'benchmark/ips'

Benchmark.ips do |x|
  x.report("[]=") do
    hash = {}
    hash[:x] = 1
  end

  x.report("store") do
    hash = {}
    hash.store(:x, 1)
  end

  x.compare!
end
```

```
Warming up --------------------------------------
                 []=    86.698k i/100ms
               store    84.689k i/100ms
Calculating -------------------------------------
                 []=      1.600M (±19.4%) i/s -      7.716M
               store      1.195M (±21.9%) i/s -      5.759M

Comparison:
                 []=:  1599601.9 i/s
               store:  1195414.2 i/s - same-ish: difference falls within error
```

Your benchmark shows that one is faster because it is doing an unnecessary delete.