@BlakeMesdag 2016-04-13
Ah you're right, very close in the margin on further testing. I had a benchmark earlier that was showing a not-insignificant difference but could be due to process differences on my laptop.